+++ packages/expo/ios/AppDelegates/ExpoAppDelegate.swift
@@ -1,13 +1,12 @@
 import Dispatch
 import Foundation
 import React_RCTAppDelegate
+import ExpoModulesCore
 
 // TODO(vonovak,20250107) - Remove the if expression when we drop SDK 52 / RN 76 support
 #if canImport(ReactAppDependencyProvider)
 import ReactAppDependencyProvider
 #endif
-
-var subscribers = [ExpoAppDelegateSubscriberProtocol]()
 
 /**
  Allows classes extending `ExpoAppDelegateSubscriber` to hook into project's app delegate
@@ -32,7 +31,7 @@ open class ExpoAppDelegate: ExpoAppInstance {
     _ application: UIApplication,
     willFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]? = nil
   ) -> Bool {
-    let parsedSubscribers = subscribers.filter {
+    let parsedSubscribers = ExpoAppDelegateSubscriberRepository.getAllSubscribers().filter {
       $0.responds(to: #selector(application(_:willFinishLaunchingWithOptions:)))
     }
 
@@ -59,7 +58,7 @@ open class ExpoAppDelegate: ExpoAppInstance {
       super.application(application, didFinishLaunchingWithOptions: launchOptions)
     }
 
-    subscribers.forEach { subscriber in
+    ExpoAppDelegateSubscriberRepository.getAllSubscribers().forEach { subscriber in
       // Subscriber result is ignored as it doesn't matter if any subscriber handled the incoming URL – we always return `true` anyway.
       _ = subscriber.application?(application, didFinishLaunchingWithOptions: launchOptions)
     }
@@ -72,25 +71,25 @@ open class ExpoAppDelegate: ExpoAppInstance {
 
   @objc
   open override func applicationDidBecomeActive(_ application: UIApplication) {
-    subscribers.forEach { $0.applicationDidBecomeActive?(application) }
+    ExpoAppDelegateSubscriberRepository.getAllSubscribers().forEach { $0.applicationDidBecomeActive?(application) }
   }
 
   @objc
   open override func applicationWillResignActive(_ application: UIApplication) {
-    subscribers.forEach { $0.applicationWillResignActive?(application) }
+    ExpoAppDelegateSubscriberRepository.getAllSubscribers().forEach { $0.applicationWillResignActive?(application) }
   }
 
   @objc
   open override func applicationDidEnterBackground(_ application: UIApplication) {
-    subscribers.forEach { $0.applicationDidEnterBackground?(application) }
+    ExpoAppDelegateSubscriberRepository.getAllSubscribers().forEach { $0.applicationDidEnterBackground?(application) }
   }
 
   open override func applicationWillEnterForeground(_ application: UIApplication) {
-    subscribers.forEach { $0.applicationWillEnterForeground?(application) }
+    ExpoAppDelegateSubscriberRepository.getAllSubscribers().forEach { $0.applicationWillEnterForeground?(application) }
   }
 
   open override func applicationWillTerminate(_ application: UIApplication) {
-    subscribers.forEach { $0.applicationWillTerminate?(application) }
+    ExpoAppDelegateSubscriberRepository.getAllSubscribers().forEach { $0.applicationWillTerminate?(application) }
   }
 
   // TODO: - Responding to Environment Changes
@@ -105,7 +104,7 @@ open class ExpoAppDelegate: ExpoAppInstance {
     completionHandler: @escaping () -> Void
   ) {
     let selector = #selector(application(_:handleEventsForBackgroundURLSession:completionHandler:))
-    let subs = subscribers.filter { $0.responds(to: selector) }
+    let subs = ExpoAppDelegateSubscriberRepository.getAllSubscribers().filter { $0.responds(to: selector) }
     var subscribersLeft = subs.count
     let dispatchQueue = DispatchQueue(label: "expo.application.handleBackgroundEvents")
 
@@ -131,11 +130,11 @@ open class ExpoAppDelegate: ExpoAppInstance {
   // MARK: - Handling Remote Notification Registration
 
   open override func application(_ application: UIApplication, didRegisterForRemoteNotificationsWithDeviceToken deviceToken: Data) {
-    subscribers.forEach { $0.application?(application, didRegisterForRemoteNotificationsWithDeviceToken: deviceToken) }
+    ExpoAppDelegateSubscriberRepository.getAllSubscribers().forEach { $0.application?(application, didRegisterForRemoteNotificationsWithDeviceToken: deviceToken) }
   }
 
   open override func application(_ application: UIApplication, didFailToRegisterForRemoteNotificationsWithError error: Error) {
-    subscribers.forEach { $0.application?(application, didFailToRegisterForRemoteNotificationsWithError: error) }
+    ExpoAppDelegateSubscriberRepository.getAllSubscribers().forEach { $0.application?(application, didFailToRegisterForRemoteNotificationsWithError: error) }
   }
 
   open override func application(
@@ -144,7 +143,7 @@ open class ExpoAppDelegate: ExpoAppInstance {
     fetchCompletionHandler completionHandler: @escaping (UIBackgroundFetchResult) -> Void
   ) {
     let selector = #selector(application(_:didReceiveRemoteNotification:fetchCompletionHandler:))
-    let subs = subscribers.filter { $0.responds(to: selector) }
+    let subs = ExpoAppDelegateSubscriberRepository.getAllSubscribers().filter { $0.responds(to: selector) }
     var subscribersLeft = subs.count
     let dispatchQueue = DispatchQueue(label: "expo.application.remoteNotification", qos: .userInteractive)
     var failedCount = 0
@@ -184,7 +183,7 @@ open class ExpoAppDelegate: ExpoAppInstance {
   // MARK: - Continuing User Activity and Handling Quick Actions
 
   open override func application(_ application: UIApplication, willContinueUserActivityWithType userActivityType: String) -> Bool {
-    return subscribers.reduce(false) { result, subscriber in
+    return ExpoAppDelegateSubscriberRepository.getAllSubscribers().reduce(false) { result, subscriber in
       return subscriber.application?(application, willContinueUserActivityWithType: userActivityType) ?? false || result
     }
   }
@@ -195,7 +194,7 @@ open class ExpoAppDelegate: ExpoAppInstance {
     restorationHandler: @escaping ([UIUserActivityRestoring]?) -> Void
   ) -> Bool {
     let selector = #selector(application(_:continue:restorationHandler:))
-    let subs = subscribers.filter { $0.responds(to: selector) }
+    let subs = ExpoAppDelegateSubscriberRepository.getAllSubscribers().filter { $0.responds(to: selector) }
     var subscribersLeft = subs.count
     let dispatchQueue = DispatchQueue(label: "expo.application.continueUserActivity", qos: .userInteractive)
     var allRestorableObjects = [UIUserActivityRestoring]()
@@ -220,11 +219,11 @@ open class ExpoAppDelegate: ExpoAppInstance {
   }
 
   open override func application(_ application: UIApplication, didUpdate userActivity: NSUserActivity) {
-    return subscribers.forEach { $0.application?(application, didUpdate: userActivity) }
+    return ExpoAppDelegateSubscriberRepository.getAllSubscribers().forEach { $0.application?(application, didUpdate: userActivity) }
   }
 
   open override func application(_ application: UIApplication, didFailToContinueUserActivityWithType userActivityType: String, error: Error) {
-    return subscribers.forEach {
+    return ExpoAppDelegateSubscriberRepository.getAllSubscribers().forEach {
       $0.application?(application, didFailToContinueUserActivityWithType: userActivityType, error: error)
     }
   }
@@ -236,7 +235,7 @@ open class ExpoAppDelegate: ExpoAppInstance {
     completionHandler: @escaping (Bool) -> Void
   ) {
     let selector = #selector(application(_:performActionFor:completionHandler:))
-    let subs = subscribers.filter { $0.responds(to: selector) }
+    let subs = ExpoAppDelegateSubscriberRepository.getAllSubscribers().filter { $0.responds(to: selector) }
     var subscribersLeft = subs.count
     var result: Bool = false
     let dispatchQueue = DispatchQueue(label: "expo.application.performAction", qos: .userInteractive)
@@ -269,7 +268,7 @@ open class ExpoAppDelegate: ExpoAppInstance {
     performFetchWithCompletionHandler completionHandler: @escaping (UIBackgroundFetchResult) -> Void
   ) {
     let selector = #selector(application(_:performFetchWithCompletionHandler:))
-    let subs = subscribers.filter { $0.responds(to: selector) }
+    let subs = ExpoAppDelegateSubscriberRepository.getAllSubscribers().filter { $0.responds(to: selector) }
     var subscribersLeft = subs.count
     let dispatchQueue = DispatchQueue(label: "expo.application.performFetch", qos: .userInteractive)
     var failedCount = 0
@@ -313,7 +312,7 @@ open class ExpoAppDelegate: ExpoAppInstance {
   // MARK: - Opening a URL-Specified Resource
 
   open override func application(_ app: UIApplication, open url: URL, options: [UIApplication.OpenURLOptionsKey: Any] = [:]) -> Bool {
-    return subscribers.reduce(false) { result, subscriber in
+    return ExpoAppDelegateSubscriberRepository.getAllSubscribers().reduce(false) { result, subscriber in
       return subscriber.application?(app, open: url, options: options) ?? false || result
     }
   }
@@ -336,7 +335,7 @@ open class ExpoAppDelegate: ExpoAppInstance {
     let universalOrientationMask = allowedOrientations(for: .unspecified)
     let infoPlistOrientations = deviceOrientationMask.isEmpty ? universalOrientationMask : deviceOrientationMask
 
-    let parsedSubscribers = subscribers.filter {
+    let parsedSubscribers = ExpoAppDelegateSubscriberRepository.getAllSubscribers().filter {
       $0.responds(to: #selector(application(_:supportedInterfaceOrientationsFor:)))
     }
 
@@ -357,33 +356,7 @@ open class ExpoAppDelegate: ExpoAppInstance {
 
   @objc
   open override func customize(_ rootView: RCTRootView) {
-    subscribers.forEach { $0.customizeRootView?(rootView) }
-  }
-
-  // MARK: - Statics
-
-  @objc
-  public static func registerSubscribersFrom(modulesProvider: ModulesProvider) {
-    modulesProvider.getAppDelegateSubscribers().forEach { subscriberType in
-      registerSubscriber(subscriberType.init())
-    }
-  }
-
-  @objc
-  public static func registerSubscriber(_ subscriber: ExpoAppDelegateSubscriberProtocol) {
-    if subscribers.contains(where: { $0 === subscriber }) {
-      fatalError("Given app delegate subscriber `\(String(describing: subscriber))` is already registered.")
-    }
-    subscribers.append(subscriber)
-  }
-
-  @objc
-  public static func getSubscriber(_ name: String) -> ExpoAppDelegateSubscriberProtocol? {
-    return subscribers.first { String(describing: $0) == name }
-  }
-
-  public static func getSubscriberOfType<Subscriber>(_ type: Subscriber.Type) -> Subscriber? {
-    return subscribers.first { $0 is Subscriber } as? Subscriber
+    ExpoAppDelegateSubscriberRepository.getAllSubscribers().forEach { $0.customizeRootView?(rootView) }
   }
 }