+++ packages/expo/CHANGELOG.md
@@ -25,6 +25,7 @@
 - Added a debug only warning and style for zero height DOM components. ([#34375](https://github.com/expo/expo/pull/34375) by [@kudo](https://github.com/kudo))
 - [Android] Started using expo modules gradle plugin. ([#34176](https://github.com/expo/expo/pull/34176) by [@lukmccall](https://github.com/lukmccall))
 - [apple] Migrate remaining `expo-module.config.json` to unified platform syntax. ([#34445](https://github.com/expo/expo/pull/34445) by [@reichhartd](https://github.com/reichhartd))
+- [apple] Move `AppDelegate` integration from `expo-modules-core` to `expo` package.
 
 ## 52.0.23 - 2024-12-24